+++ server/package.json
@@ -7,7 +7,7 @@
     "dev": "NODE_ENV=development nodemon --config './nodemon.json' app.js",
     "start": "NODE_ENV=production node app.js",
     "test": "NODE_ENV=test && mocha src/**/*.test.js --exit",
-    "test-watch": "export NODE_ENV=test && mocha src/**/*.test.js"
+    "test-watch": "NODE_ENV=test mocha src/**/*.test.js --watch"
   },
   "repository": {
     "type": "git",
@@ -22,22 +22,22 @@
   "dependencies": {
     "axios": "^0.17.1",
     "bcryptjs": "^2.4.3",
-    "chalk": "^2.3.0",
-    "cloudinary": "^1.9.1",
+    "chalk": "^2.3.2",
+    "cloudinary": "^1.11.0",
     "dotenv": "^4.0.0",
-    "express": "^4.16.2",
-    "joi": "^13.1.0",
-    "jsonwebtoken": "^8.1.0",
-    "mongoose": "^5.0.0",
+    "express": "^4.16.3",
+    "joi": "^13.1.2",
+    "jsonwebtoken": "^8.2.0",
+    "mongoose": "^5.0.12",
     "mongoose-sequence": "^4.0.1",
     "mongoose-timestamp": "^0.6.0",
     "morgan": "^1.9.0",
     "multer": "^1.3.0",
-    "nodemailer": "^4.4.1"
+    "nodemailer": "^4.6.4"
   },
   "devDependencies": {
     "chai": "^4.1.2",
     "chai-http": "^3.0.0",
-    "mocha": "^5.0.1"
+    "mocha": "^5.0.5"
   }
 }
