+++ server/src/routes/auth/auth.test.js
@@ -1,6 +1,6 @@
 const chai = require('chai');
 const chaiHttp = require('chai-http');
-const expect = require('chai').expect;
+const { expect } = require('chai');
 
 const app = require('../../../app');
 
@@ -8,28 +8,16 @@ chai.use(chaiHttp);
 
 describe('POST /request/verification', () => {
 
-  // it('should send out a verification email to a given email address', () => {
-  //
-  //   chai
-  //     .request(app)
-  //     .post('/request/verification')
-  //     .send({
-  //       'email': 'ckboardtoronto@gmail.com'
-  //     });
-  //
-  // });
-
   it('should reject an invalid email address', (done) => {
 
     chai
       .request(app)
       .post('/api/auth/request/verification')
       .send({
-        'email': 'email@address.com'
+        'email': 'emailaddress.com'
       })
-      .end(function (err, res) {
-        expect(null).to.be.null;
-        // expect(err).to.have.status(400);
+      .end((err, res) => {
+        expect(res).to.have.status(400);
         done();
       });
